+++ src/main/java/com/example/db/repo/EmployeeRepo.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface EmployeeRepo extends JpaRepository<Employee, Long> {
+public interface EmployeeRepo extends JpaRepository<Employee, Integer> {
 
     Optional<Employee> findEmployeeByFirstNameAndPatherNameAndLastName(String firstName, String patherName, String lastName);
 }
